test resetPassword component
