+++ src/views/ResetPassword/ResetPassword.js
@@ -1,6 +1,6 @@
 import React from 'react';
 // eslint-disable-next-line import/no-named-as-default
-import ResetPasswordCard from '../../components/ResetPassword/ResetPasswordCard';
+import ResetPasswordCard from '../../components/ResetPassword/ResetPassword';
 
 const ResetPassword = () => (
   <ResetPasswordCard />
